+++ Khoaluan/Khoaluan/Models/Item.cs
@@ -7,8 +7,8 @@ namespace Khoaluan.Models
         public string Name { get; set; }
         public string Image { get; set; }
         public int ProductId { get; set; }
-        public int MaxPrice { get; set; }
-        public int MinPrice { get; set; }
+        public decimal MaxPrice { get; set; }
+        public decimal MinPrice { get; set; }
         public Product Product { get; set; }
         public List<Inventory> Inventories { get; set; }
         public List<Market> Markets { get; set; }
